# Conflicts:
#	ChronoLoop/ChronoLoop/Actions/BoxSnapToControllerAction.hpp
#	ChronoLoop/ChronoLoop/Input/Controller.cpp
#	ChronoLoop/ChronoLoop/Input/VRInputManager.cpp

+++ ChronoLoop/ChronoLoop/Core/TimeManager.cpp
@@ -379,7 +379,7 @@ namespace Epoch {
 				return;
 			}
 
-			if (mtempCurSnapFrame > mTimeline->GetCurrentGameTimeIndx())
+			if ((unsigned int)mtempCurSnapFrame > mTimeline->GetCurrentGameTimeIndx())
 				return;
